chore: returned themes cofig into DocsGPTWidget component
